refactor(events) : separate EventListViewModel and EventViewModel for consistency
This will separate the user’s actions from the broader event management logic.
